+++ variants/cardputer_adv/companion_radio/ui-cardputer/screen/NewMessageScreen.h
@@ -1,0 +1,17 @@
+#include "../../CardputerMesh.h"
+#include "../CardputerUITask.h"
+#include "globals.h"
+
+class NewMessageScreen : public CardputerScreen {
+  CardputerUITask *_task;
+  char origin[62];
+  char message[MAX_MESSAGE_LENGTH];
+  unsigned long dismiss_after = 0;
+
+public:
+  NewMessageScreen(CardputerUITask *task) : _task(task) {}
+  void newMessage(const char *from, const char *msg);
+  int render(DisplayDriver &display) override;
+  bool handleInput(Keyboard::Event &e) override;
+  void poll() override;
+};